Added UI, change canvas size, and rewrote update function
Rendering the highlighted UI option is slow.
Added an extra field to appstate, "ui", which holds all relevant information.
Changed the way the tool-color value is obtained, now by checking the value of "sel" in ui.
str->icon uses list-ref for handling palette icons, altough it was explicitly stated on the project guidelines to not use such a function. Should be changed.